+++ include/wkhtmltox/pdfsettings.hh
@@ -87,6 +87,9 @@ struct DLL_PUBLIC PdfGlobal {
 
 	//! Should we use the graphics system
 	bool useGraphics;
+	
+	//! Should relative links be resolved
+	bool resolveRelativeLinks;
 
 	//! Should we orientate in landscape or portrate
 	QPrinter::Orientation orientation;