+++ src/lib/pdfsettings.hh
@@ -91,6 +91,9 @@ struct DLL_PUBLIC PdfGlobal {
 	//! Should we use the graphics system
 	bool useGraphics;
 
+	//! Should relative links be resolved or kept as-is
+	bool resolveRelativeLinks;
+
 	//! Should we orientate in landscape or portrate
 	QPrinter::Orientation orientation;