+++ CHANGELOG.md
@@ -4,6 +4,7 @@ v0.12.3 (unreleased)
 * removed support for CentOS 5 builds
 * update OpenSSL to 1.0.2a and xz to 5.2.1
 * downgrade libpng to 1.2.53
+* **#1843**: add --keep-relative-links and --resolve-relative-links to resolve relative links
 * **#2104**: renamed COPYING to LICENSE
 * **#2190**: do not depend on ICU even if it is already installed
 * **#2280**: do not allow data URIs for --header-html or --footer-html